+++ CHANGELOG.md
@@ -35,7 +35,27 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   - `OpenTelemetry.Instrumentation.EntityFrameworkCore`
     from `1.15.0-beta.1` to `1.15.1-beta.1`.
 - .NET Framework only, following packages updated
-  - `OpenTelemetry.Instrumentation.AspNet` from `1.15.1` to `1.15.2`.
+  - `Microsoft.Bcl.AsyncInterfaces` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Configuration` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Configuration.Abstractions` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Configuration.Binder` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.DependencyInjection` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.DependencyInjection.Abstractions`
+    from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Diagnostics.Abstractions` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Logging` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Logging.Abstractions` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Logging.Configuration` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Options` from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Options.ConfigurationExtensions`
+    from `10.0.2` to `10.0.7`,
+  - `Microsoft.Extensions.Primitives` from `10.0.2` to `10.0.7`,
+  - `OpenTelemetry.Instrumentation.AspNet` from `1.15.1` to `1.15.2`,
+  - `System.Diagnostics.DiagnosticSource` from `10.0.2` to `10.0.7`,
+  - `System.IO.Pipelines` from `10.0.2` to `10.0.7`,
+  - `System.Text.Encodings.Web` from `10.0.2` to `10.0.7`,
+  - `System.Text.Json` from `10.0.2` to `10.0.7`,
+  - `System.ValueTuple` from `4.6.1` to `4.6.2`.
 
 ### Deprecated